:truck: Rename CreateAccount component to CreateAccountButton (more representative name)
