+++ src/modules/Auth/components/CreateAccountButton/CreateAccountButton.js
@@ -5,9 +5,9 @@ import i18next from 'i18next';
 import { useTheme } from 'contexts/ThemeContext';
 import { P, B } from 'components/shared/toolBox/typography';
 
-import getCreateAccountStyles from './styles';
+import getCreateAccountStyles from './CreateAccountButton.styles';
 
-export default function CreateAccount({ onPress, style, opacity = 1 }) {
+export default function CreateAccountButton({ onPress, style, opacity = 1 }) {
   const { styles } = useTheme({
     styles: getCreateAccountStyles(),
   });
@@ -15,7 +15,7 @@ export default function CreateAccount({ onPress, style, opacity = 1 }) {
   return (
     <Animated.View style={[styles.container, style?.container, { opacity }]}>
       <P style={[styles.question, styles.theme.question, style?.question]}>
-        {i18next.t('Don’t have a Lisk Account?')}
+        {i18next.t('auth.register.createAccountButton.text')}
       </P>
 
       <B
@@ -23,7 +23,7 @@ export default function CreateAccount({ onPress, style, opacity = 1 }) {
         style={[styles.link, styles.theme.link, style?.link]}
         onPress={onPress}
       >
-        {i18next.t('Create it now')}
+        {i18next.t('auth.register.createAccountButton.buttonText')}
       </B>
     </Animated.View>
   );
